Fixed errors in tests

+++ crates/kodept-parse/src/lexer/mod.rs
@@ -57,7 +57,7 @@ mod tests {
         #[case] expected_rest: Option<&'static str>,
     ) {
         let data = DefaultLexer::new().parse_token(input, 0).unwrap();
-        let rest = &input[data.span.point.length..];
+        let rest = &input[data.span.point.length as usize..];
 
         assert_eq!(data.token, expected.into());
         assert_eq!(rest, expected_rest.unwrap_or(""));
